Use 64 bit hash and fix out of bounds access
There doesn't seem to be any statisticalyl significant changes in
performance. I filtered Ruby processes and ran `loop { puts :a }` in a
loop.

```
sudo bpftool prog | grep profile_cpu
```

```
2938: perf_event  name profile_cpu  tag 4554576c1d62dc61  gpl run_time_ns 136221451 run_cnt 24408 -> 5581.02 ns / run
              2268 run_cnt
          60823512 cycles        -> 26818.13 cycles / run
          42618949 instructions    -> 18791.42 instructions / run #     0.70 insns per cycle
```

```
3004: perf_event  name profile_cpu  tag 75f4ac73b948f3c0q  gpl run_time_ns 144678180 run_cnt 23700 -> 6104.56 ns / run

             2279 run_cnt
          60910573 cycles            -> 26726.89 cycles / run
          42422810 instructions       -> 18614.66 instructions / run #     0.70 insns per cycle
```

+++ bpf/unwinders/shared.h
@@ -14,8 +14,8 @@ typedef struct {
     int tgid;
     int user_stack_id;
     int kernel_stack_id;
-    int user_stack_id_dwarf_id;
-    int interpreter_stack_id;
+    u64 user_stack_id_dwarf_id;
+    u64 interpreter_stack_id;
 } stack_count_key_t;
 
 typedef struct {
@@ -47,7 +47,7 @@ struct {
 struct {
     __uint(type, BPF_MAP_TYPE_HASH);
     __uint(max_entries, MAX_STACK_COUNTS_ENTRIES);
-    __type(key, int);
+    __type(key, u64);
     __type(value, stack_trace_t);
 } interpreter_stack_traces SEC(".maps"); // TODO think about this.
